refactor(editor): move block yjs listener to store (#12140)
